Fix calculate_mobility bug with en passant
When calculating mobility, the color is switched. Because of that,
we can end up in a case where, if the original color (before the
switch) could en passant, the board will also think the current color
can en passant. Not only this wrongly increases the mobility, but
it can lead to a bad state since some places of the board assume
an en passant is a capture.